+++ src/main/java/net/jrtechs/www/App.java
@@ -1,12 +1,13 @@
 package net.jrtechs.www;
 
-import net.jrtechs.www.server.Server;
+import net.jrtechs.www.server.old.Server;
 
 /**
  * Launcher for the server
  *
  * @author Jeffery Russell 6-9-18
  */
+@Deprecated
 public class App 
 {
     public static void main( String[] args )
