+++ src/main/java/net/jrtechs/www/server/old/Client.java
@@ -1,6 +1,7 @@
-package net.jrtechs.www.server;
+package net.jrtechs.www.server.old;
 
 import net.jrtechs.www.graphDB.SteamGraph;
+import net.jrtechs.www.model.Player;
 import org.java_websocket.WebSocket;
 import org.json.JSONObject;
