+++ src/main/java/net/jrtechs/www/server/old/Server.java
@@ -1,4 +1,4 @@
-package net.jrtechs.www.server;
+package net.jrtechs.www.server.old;
 
 import org.java_websocket.WebSocket;
 import org.java_websocket.handshake.ClientHandshake;
